+++ src/components/ComponentsProvider/registry.ts
@@ -27,7 +27,7 @@ type ComponentType<T> = T extends React.ComponentType<any>
     ? React.ComponentType<React.ComponentPropsWithoutRef<T>>
     : never;
 
-export interface ComponentRegistry<T extends Registry, Entities = NonNullable<T['type']>> {
+export interface ComponentsRegistry<T extends Registry, Entities = NonNullable<T['type']>> {
     set<Id extends keyof Entities>(id: Id, entity: ComponentType<Entities[Id]>): this;
     get<Id extends keyof Entities>(id: Id): ComponentType<Entities[Id]>;
 }
